// src/app/book/shared/models/books.model.ts
export interface BookData {
  title: string;
  subtitle: string;
  isbn13: string;
  price: string;
  image: string;
  url: string;
}
export interface BookDataDetail {
  title: string;
  subtitle: string;
  authors: string;
  publisher: string;
  isbn10: string;
  isbn13: string;
  pages: number;
  year: number;
  rating: number;
  desc: string;
  price: string;
  image: string;
  url: string;
  language: string;
}